Split internal private classes into their own files, split a few other concepts into their own files
